Update to endpoints

+++ controllers/productControllers.js
@@ -1,6 +1,8 @@
 import Product from "../models/product.js";
 import Variant from "../models/productVariant.js";
 
+let existingProduct;
+
 const getAllProductController = async (req, res) => {
   const allProducts = await Product.find({});
 
@@ -61,7 +63,7 @@ const createProductController = async (req, res) => {
       const savedProduct = await newProduct.save();
       res.status(200).json({
         success: true,
-        message: "New Product get saved successfully!!",
+        message: "New Product saved successfully!!",
         savedProduct,
       });
     }
@@ -78,11 +80,20 @@ const deleteProductByIdController = async (req, res) => {
   try {
     const pId = req.params.productId;
 
+    existingProduct = await Product.findById(pId);
+
+    if (!existingProduct) {
+      return res.status(404).json({
+        success: false,
+        message: "Product was not found.",
+      });
+    }
+
     const deletedProduct = await Product.findByIdAndDelete(pId);
 
     res.status(200).json({
       success: true,
-      message: `Product with id ${pId} get deleted`,
+      message: `Product with id ${pId} is deleted`,
       deletedProduct,
     });
   } catch (error) {
@@ -96,13 +107,24 @@ const deleteProductByIdController = async (req, res) => {
 const updateProductById = async (req, res) => {
   try {
     const pId = req.params.productId;
-    const changes = { price: 120000 };
+    const changes = req.body;
     const beforeUpdateProduct = await Product.findByIdAndUpdate(pId, changes);
+
+    existingProduct = await Product.findById(pId);
+
+    if (!existingProduct) {
+      return res.status(404).json({
+        success: false,
+        message: "Product was not found.",
+      });
+    }
+
     res.status(200).json({
       success: true,
-      message: "Product get updated and below is it's older form",
+      message: "Product is updated and below is it's older form",
       beforeUpdateProduct,
     });
+
   } catch (error) {
     res.status(500).json({
       success: false,
